:art: Stop using cryptic symbols and other stuff :fire:
