+++ examples/vararg.gw
@@ -1,15 +1,14 @@
 fun void test(...) {
   int i;
-  vararg.start;
-
-	if(i == 0)
-		<<< vararg $ int >>>;
-	else if(i == 1)
-		<<< vararg $ float >>>;
-	else if(i == 2)
-		<<< vararg $ Object >>>;
-  	i++;
-  vararg.end;
+  varloop vararg {
+  	if(i == 0)
+	  	<<< vararg $ int >>>;
+  	else if(i == 1)
+	  	<<< vararg $ float >>>;
+  	else if(i == 2)
+	  	<<< vararg $ Object >>>;
+  	  i++;
+  }
 }
 test(1);
 test(1, 2.3, null);
