+++ include/import/oper.h
@@ -7,6 +7,9 @@ ANN m_int gwi_oper_emi(const Gwi gwi, const opem);
 ANN2(1) m_int gwi_oper_end(const Gwi gwi, const m_str op, const f_instr f);
 ANN m_int gwi_oper_cond(const Gwi, const m_str,  const f_instr, const f_instr);
 
+//! use this with one if the right side operator's to emit_var is set
+ANN m_int gwi_oper_var(const Gwi gwi, const m_bool offset);
+
 #define _EQUALITY_OPER(sz, sign)                                 \
   POP_REG(shred, sz*2 - SZ_INT);                                 \
   *(m_uint*)REG(-SZ_INT) = sign                                  \
