+++ include/lang_private.h
@@ -11,4 +11,6 @@ ANN m_bool import_array(const Gwi gwi);
 ANN m_bool import_ptr(const Gwi gwi);
 ANN m_bool import_func(const Gwi gwi);
 ANN m_bool import_modules(const Gwi gwi);
+ANN m_bool import_object_op(const Gwi gwi);
+ANN m_bool import_values(const Gwi gwi);
 #endif
