+++ src/import/oper.c
@@ -16,15 +16,13 @@
 #include "mpool.h"
 #include "specialid.h"
 
-ANN static Type _get_type(const Gwi gwi, const m_str str) {
-  if(str == (m_str)OP_ANY_TYPE)
+ANN static Type _get_type(const Gwi gwi, const m_str s) {
+  if(s == (m_str)OP_ANY_TYPE)
     return OP_ANY_TYPE;
-  struct array_checker ck = { .str=str };
-  DECL_OO(const ID_List, list, = ck2list(gwi, &ck))
-  const Type t = find_type(gwi->gwion->env, list);
-  if(list)
-    free_id_list(gwi->gwion->mp, list);
-  return t ? (ck.depth ? array_type(gwi->gwion->env, t, ck.depth) : t) : NULL;
+  Type_Decl *td = str2decl(gwi, s);
+  const Type t = known_type(gwi->gwion->env, td);
+  free_type_decl(gwi->gwion->mp, td);
+  return t;
 }
 
 ANN2(1) static inline Type get_type(const Gwi gwi, const m_str str) {
@@ -36,8 +34,9 @@ ANN2(1,2) static int import_op(const Gwi gwi, const struct OperCK* op,
   const Type lhs = get_type(gwi, op->lhs),
              rhs = get_type(gwi, op->rhs),
              ret = get_type(gwi, op->ret);
-  const struct Op_Import opi = { lhs, rhs, ret,
-    op->ck, op->em, (uintptr_t)f, gwi->loc, op->sym };
+  const struct Op_Func opfunc = { .ck=op->ck, .em=op->em };
+  const struct Op_Import opi = { .lhs=lhs, .rhs=rhs, .ret=ret,
+    .func=&opfunc, .data=(uintptr_t)f, .pos=gwi->loc, .op=op->sym, .emit_var=op->emit_var };
   return add_op(gwi->gwion, &opi);
 }
 
@@ -52,6 +51,11 @@ ANN2(1) m_int gwi_oper_ini(const Gwi gwi, const restrict m_str l,
 
 ANN m_int gwi_oper_add(const Gwi gwi, Type (*ck)(Env, void*, m_bool*)) {
   gwi->oper->ck = ck;
+  return GW_OK;
+}
+
+ANN m_int gwi_oper_var(const Gwi gwi, const m_bool offset) {
+  gwi->oper->emit_var = offset;
   return GW_OK;
 }
 
@@ -70,7 +74,7 @@ ANN m_int gwi_oper_end(const Gwi gwi, const m_str op, const f_instr f) {
 
 ANN m_int gwi_oper_cond(const Gwi gwi, const m_str type,
   const f_instr f1, const f_instr f2) {
-  GWI_BB(gwi_oper_ini(gwi, NULL, type, "int"))
+  GWI_BB(gwi_oper_ini(gwi, NULL, type, "bool"))
   GWI_BB(gwi_oper_end(gwi, "@conditionnal", f1))
   GWI_BB(gwi_oper_end(gwi, "@unconditionnal", f2))
   return GW_OK;
