+++ src/lib/event.c
@@ -24,8 +24,6 @@ static OP_CHECK(opck_eventwait) {
 static INSTR(EventWait) {
   POP_REG(shred, SZ_FLOAT);
   const M_Object event = *(M_Object*)REG(-SZ_INT);
-  if(!event)
-    Except(shred, "NullEventWait");
   shreduler_remove(shred->tick->shreduler, shred, 0);
   const Vector v = EV_SHREDS(event);
   vector_add(v, (vtype)shred);
@@ -66,7 +64,8 @@ GWION_IMPORT(event) {
   GWI_BB(gwi_func_ini(gwi, "void", "broadcast"))
   GWI_BB(gwi_func_end(gwi, event_broadcast, ae_flag_none))
   GWI_BB(gwi_class_end(gwi))
-  GWI_BB(gwi_oper_ini(gwi, "Event", "@now", "int"))
+  GWI_BB(gwi_oper_ini(gwi, "nonnull Event", "@now", "int"))
+  GWI_BB(gwi_oper_var(gwi, -SZ_FLOAT))
   _CHECK_OP("=>", eventwait, EventWait)
   return GW_OK;
 }
