+++ src/lib/ptr.c
@@ -20,9 +20,14 @@ static OP_CHECK(opck_ptr_assign) {
     ERR_N(exp_self(bin)->pos, _("left side operand is constant"));
   bin->lhs->emit_var = 1;
   Type t = bin->lhs->type;
-  do if(!strcmp(t->name, get_type_name(env, bin->rhs->type->name, 1)))
-    return bin->lhs->type;
-  while((t = t->e->parent));
+  do {
+    Type u = bin->rhs->type;
+    do {
+      const m_str str = get_type_name(env, u->name, 1);
+      if(str && !strcmp(t->name, str))
+        return bin->lhs->type;
+    } while((u = u->e->parent));
+  } while((t = t->e->parent));
   return env->gwion->type[et_null];
 }
 
@@ -32,32 +37,24 @@ static INSTR(instr_ptr_assign) {
   *(m_uint**)o->data = *(m_uint**)REG(-SZ_INT);
 }
 
-static OP_EMIT(opem_ptr_assign) {
-  const Exp_Binary* bin = (Exp_Binary*)data;
-  if(!GET_FLAG(bin->rhs->type, nonnull))
-    emit_add_instr(emit, GWOP_EXCEPT);
-  return emit_add_instr(emit, instr_ptr_assign);
-}
-
 static OP_CHECK(opck_ptr_deref) {
   const Exp_Unary* unary = (Exp_Unary*)data;
-  return exp_self(unary)->type = nspc_lookup_type1(env->curr, insert_symbol(get_type_name(env, unary->exp->type->name, 1)));
+  DECL_ON(const m_str, str, = get_type_name(env, unary->exp->type->name, 1))
+  return exp_self(unary)->type = nspc_lookup_type1(env->curr, insert_symbol(str));
 }
 
 static OP_CHECK(opck_ptr_cast) {
   const Exp_Cast* cast = (Exp_Cast*)data;
   DECL_ON(const Type, t, = type_decl_resolve(env, cast->td))
-  if(!GET_FLAG(t, check)) {
-    assert(t->e->def);
+  if(!GET_FLAG(t, check))
     CHECK_BN(traverse_class_def(env, t->e->def))
-  }
   const Type to = known_type(env, cast->td->types->td);
   if(isa(cast->exp->type, to) > 0)
     return t;
   ERR_N(exp_self(cast)->pos, "invalid pointer cast")
 }
 
-static OP_CHECK(opck_implicit_ptr) {
+static OP_CHECK(opck_ptr_implicit) {
   const struct Implicit* imp = (struct Implicit*)data;
   const Exp e = imp->e;
   if(!strcmp(get_type_name(env, imp->t->name, 1), e->type->name)) {
@@ -65,6 +62,8 @@ static OP_CHECK(opck_implicit_ptr) {
       ERR_N(e->pos, _("can't cast constant to Ptr"));
     e->cast_to = imp->t;
     e->emit_var = 1;
+    if(!GET_FLAG(imp->t, check))
+      CHECK_BN(traverse_class_def(env, imp->t->e->def))
     return imp->t;
   }
   return NULL;
@@ -83,18 +82,25 @@ static INSTR(instr_ptr_deref) {
 }
 
 static INSTR(Cast2Ptr) {
-  const M_Object o = new_object(shred->info->mp, shred, shred->info->vm->gwion->type[et_ptr]);
+  const M_Object o = new_object(shred->info->mp, shred, (Type)instr->m_val);
   *(m_uint**)o->data = *(m_uint**)REG(-SZ_INT);
   *(M_Object*)REG(-SZ_INT) = o;
 }
-/*
+
 static OP_EMIT(opem_ptr_cast) {
   const Exp_Cast* cast = (Exp_Cast*)data;
   const Instr instr = emit_add_instr(emit, Cast2Ptr);
   instr->m_val = (m_uint)exp_self(cast)->type;
-  return GW_OK;
+  return instr;
 }
-*/
+
+static OP_EMIT(opem_ptr_implicit) {
+  const struct Implicit* imp = (struct Implicit*)data;
+  const Instr instr = emit_add_instr(emit, Cast2Ptr);
+  instr->m_val = (m_uint)imp->t;
+  return instr;
+}
+
 static OP_EMIT(opem_ptr_deref) {
   const Exp_Unary* unary = (Exp_Unary*)data;
   const Instr instr = emit_add_instr(emit, instr_ptr_deref);
@@ -109,17 +115,17 @@ GWION_IMPORT(ptr) {
   GWI_BB(gwi_item_ini(gwi, "@internal", "@val"))
   GWI_BB(gwi_item_end(gwi, 0, NULL))
   GWI_BB(gwi_class_end(gwi))
-  t_ptr->nspc->info->offset = SZ_INT;
-  GWI_BB(gwi_oper_ini(gwi, (m_str)OP_ANY_TYPE, "Ptr", NULL))
+  SET_FLAG(t_ptr, unary);
+  GWI_BB(gwi_oper_ini(gwi, (m_str)OP_ANY_TYPE, "nonnull Ptr", NULL))
   GWI_BB(gwi_oper_add(gwi, opck_ptr_assign))
-  GWI_BB(gwi_oper_emi(gwi, opem_ptr_assign))
   GWI_BB(gwi_oper_end(gwi, ":=>", instr_ptr_assign))
-  GWI_BB(gwi_oper_add(gwi, opck_implicit_ptr))
+  GWI_BB(gwi_oper_add(gwi, opck_ptr_implicit))
+  GWI_BB(gwi_oper_emi(gwi, opem_ptr_implicit))
   GWI_BB(gwi_oper_end(gwi, "@implicit", Cast2Ptr))
-  GWI_BB(gwi_oper_ini(gwi, (m_str)OP_ANY_TYPE, "Ptr", NULL))
   GWI_BB(gwi_oper_add(gwi, opck_ptr_cast))
+  GWI_BB(gwi_oper_emi(gwi, opem_ptr_cast))
   GWI_BB(gwi_oper_end(gwi, "$", Cast2Ptr))
-  GWI_BB(gwi_oper_ini(gwi, NULL, "Ptr", NULL))
+  GWI_BB(gwi_oper_ini(gwi, NULL, "nonnull Ptr", NULL))
   GWI_BB(gwi_oper_add(gwi, opck_ptr_deref))
   GWI_BB(gwi_oper_emi(gwi, opem_ptr_deref))
   GWI_BB(gwi_oper_end(gwi, "*", instr_ptr_deref))
