+++ tests/error/ev.gw
@@ -1,3 +1,3 @@
-#! [contains] NullEventWait
+#! [contains] NullPtrException
 Event ref e;
 e => now;
