--- tests/error/vararg_after.gw
@@ -1,6 +1,0 @@
-#! [contains] used after
-fun void test(...) {
-  vararg.start;
-  vararg.end;
-  vararg $ int;
-}
