+++ tests/error/vararg_outside_varloop.gw
@@ -1,0 +1,5 @@
+#! [contains] outside varloop
+fun void test(...) {
+  <<<vararg $ int >>>;
+}
+new Object => test;
