+++ tests/error/varloop_not_vararg.gw
@@ -1,0 +1,4 @@
+#! varloop expression type must be
+varloop 1 {
+
+}
