+++ tests/import/gwi_oper_emit_var.gw
@@ -1,0 +1,5 @@
+1234.5 => float f;
+PtrExt ref ptr;
+<<< ptr >>>;
+ptr <=: f;
+<<< f >>>;
