+++ tests/sh/import.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# [test] #74
+# [test] #75
 
 n=0
 [ "$1" ] && n="$1"
